SINGA-247 Add windows support for singa
Fix some warnings, most of which are type transferring.
Add a target named "singa_static" to generate static libs.

+++ src/model/layer/pooling.cc
@@ -167,7 +167,7 @@ void Pooling::ForwardMaxPooling(const float* bottom, const int num,
               const int index = h * width + w;
               if (bottom[index] > top[top_index]) {
                 top[top_index] = bottom[index];
-                mask[top_index] = index;
+                mask[top_index] = (float) index;
               }
             }
           }
